backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3
celery[redis]==5.3.6
redis==5.0.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
# PyTorch
torch==2.2.2
joblib==1.4.0
httpx==0.27.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.5.0
python-dotenv==1.0.1
